+++ ChatClient/Mail/MailTypes.hpp
